Most pet stuff are fixed They are buyable and can be spawned. And will stay when you cc :D
git-svn-id: https://svn6.assembla.com/svn/MoopleDEV@61 59296cf3-ab24-4890-a14e-c0dede3ece5d

+++ src/net/channel/handler/MovePetHandler.java
@@ -38,7 +38,7 @@ public final class MovePetHandler extends AbstractMovementPacketHandler {
             return;
         }
         MapleCharacter player = c.getPlayer();
-        byte slot = player.getPetIndex(petId);
+        int slot = player.getPetIndex(petId);
         if (slot == -1) {
             return;
         }
